requirements.txt
customtkinter==5.2.1
pycryptodome==3.19.0 